+++ include/nstd/events/def.h
@@ -6,4 +6,22 @@
 /// Represents a button ID.
 typedef NSTDUInt32 NSTDButtonID;
 
+/// Represents a window's position.
+typedef struct
+{
+    /// The position on the x-axis.
+    NSTDInt32 x;
+    /// The position on the y-axis.
+    NSTDInt32 y;
+} NSTDWindowPosition;
+
+/// Represents a window's size.
+typedef struct
+{
+    /// The width.
+    NSTDUInt32 width;
+    /// The height.
+    NSTDUInt32 height;
+} NSTDWindowSize;
+
 #endif