+++ include/nstd/gui.h
@@ -1,6 +1,5 @@
 #ifndef NSTD_GUI_H_INCLUDED
 #define NSTD_GUI_H_INCLUDED
-#include "gui/def.h"
 #include "gui/display.h"
 #include "gui/window.h"
 #include "nstd.h"