+++ src/gui.rs
@@ -1,4 +1,3 @@
 //! Graphical user interface library.
-pub mod def;
 pub mod display;
 pub mod window;